@@ -1,18 +1,18 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace MyFirstApplication.Models;
 
 public class Network
 {
-    [JsonProperty("id")]
+    [JsonPropertyName("id")]
     public int Id { get; set; }
 
-    [JsonProperty("name")]
-    public string Name { get; set; }
+    [JsonPropertyName("name")]
+    public string? Name { get; set; }
 
-    [JsonProperty("country")]
-    public Country Country { get; set; }
+    [JsonPropertyName("country")]
+    public Country? Country { get; set; }
 
-    [JsonProperty("officialSite")]
-    public string OfficialSite { get; set; }
+    [JsonPropertyName("officialSite")]
+    public string? OfficialSite { get; set; }
 }
