@@ -1,9 +1,9 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace MyFirstApplication.Models;
 
 public class Self
 {
-    [JsonProperty("href")]
+    [JsonPropertyName("href")]
     public string? Href { get; set; }
 }
