@@ -1,15 +1,15 @@
-﻿@model List<MyFirstApplication.Models.TvShowViewModel>
+﻿@model MyFirstApplication.Models.TvShowViewModel
 
 <div class="text-center">
     <h1 class="display-4">Welcome</h1>
     <p><a>Display List Of Shows</a>.</p>
 
+
     @{
-        var totalPages = (int)ViewData["TotalPages"];
         var routeName = ViewContext.ActionDescriptor.AttributeRouteInfo.Name;
         var paginationVewModel = new PaginationViewModel
             {
-                TotalPages = totalPages,
+                TotalPages = Model.TotalPages,
                 RouteName = routeName
             };
     }
@@ -21,7 +21,7 @@
     <table>
         @{
             int itemsPerRow = 4;
-            int totalItems = Model.Count;        
+            int totalItems = Model.TvShows.Count;        
 
             for (int i = 0; i < totalItems; i++)
             {
@@ -31,14 +31,14 @@
                 }
 
                 <td>
-                    <a href="@Model[i].URL">
-                        <img src="@Model[i].ImageUrl" alt="@Model[i].Name">
+                    <a href="@Model.TvShows[i].URL">
+                        <img src="@Model.TvShows[i].ImageUrl" alt="@Model.TvShows[i].Name">
                     </a>
                     <div>
-                        <a href="@Model[i].URL">
-                            <h4 style="color:black;"><b>@Model[i].Name</b></h4>
+                        <a href="@Model.TvShows[i].URL">
+                            <h4 style="color:black;"><b>@Model.TvShows[i].Name</b></h4>
                         </a>
-                        <p>Rating @Model[i].Rating?.Average</p>
+                        <p>Rating @Model.TvShows[i].Rating?.Average</p>
                     </div>
                 </td>
                 if (i % itemsPerRow == itemsPerRow - 1 || i == totalItems - 1) // Close the row
